% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCplot.R
\name{CCplot}
\alias{CCplot}
\title{Concordance Correlation Plot}
\usage{
CCplot(method1, method2, Ptype = "None", metrics = FALSE, xlabel = "",
  ylabel = "", title = "", subtitle = "", xrange = NULL,
  yrange = NULL, MArange = c(-3.5, 5.5))
}
\arguments{
\item{method1}{measurements obtained in batch 1 or using method 1}

\item{method2}{measurements obtained in batch 2 or using method 2}

\item{Ptype}{type of plot to be outputted c("scatter", "MAplot")}

\item{metrics}{if return metrics is set to \code{TRUE}, returns Rc, Ca and R2}

\item{xlabel}{label for x axis}

\item{ylabel}{label for y axis}

\item{title}{title for the main plot}

\item{subtitle}{subtitle of plot}

\item{xrange}{range of x axis}

\item{yrange}{range of y axis}

\item{MArange}{MA range}
}
\value{
Either a scatterplot or MA plot showing concordance correlation.
}
\description{
Plotting function for reliability measure.
}
\examples{
# Simulate normally distributed data
set.seed(12)
a1 <- rnorm(20) + 2
a2 <- a1 + rnorm(20, 0, 0.15)
a3 <- a1 + rnorm(20, 0, 0.15) + 1.4
a4 <- 1.5 * a1 + rnorm(20, 0, 0.15)
a5 <- 1.3 * a1 + rnorm(20, 0, 0.15) + 1
a6 <- a1 + rnorm(20, 0, 0.8)
par(mfrow = c(3, 2), mar = c(5.1, 4.1, 1.5, 1.5))

# Scatterplots
CCplot(a1, a1, Ptype = "scatter", "X", "Y", "Perfect Agreement", subtitle = letters[1])
CCplot(a1, a2, Ptype = "scatter", "X", "Y", "Very Good Agreement", subtitle = letters[2])
CCplot(a1, a3, Ptype = "scatter", "X", "Y", "Location Shift", subtitle = letters[3])
CCplot(a1, a4, Ptype = "scatter", "X", "Y", "Scale Shift", subtitle = letters[4])
CCplot(a1, a5, Ptype = "scatter", "X", "Y", "Location and Scale Shift", subtitle = letters[5])
CCplot(a1, a6, Ptype = "scatter", "X", "Y", "Measurement Error", subtitle = letters[6])

# MAplots
CCplot(a1, a1, Ptype = "MAplot", "X", "Y", "Perfect Agreement", subtitle = letters[1])
CCplot(a1, a2, Ptype = "MAplot", "X", "Y", "Very Good Agreement", subtitle = letters[2])
CCplot(a1, a3, Ptype = "MAplot", "X", "Y", "Location Shift", subtitle = letters[3])
CCplot(a1, a4, Ptype = "MAplot", "X", "Y", "Scale Shift", subtitle = letters[4])
CCplot(a1, a5, Ptype = "MAplot", "X", "Y", "Location and Scale Shift", subtitle = letters[5])
CCplot(a1, a6, Ptype = "MAplot", "X", "Y", "Measurement Error", subtitle = letters[6])
}
\author{
Aline Talhouk

Aline Talhouk
}

