% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanostring_YL_functions.R
\name{create_summary_table}
\alias{create_summary_table}
\title{create summary table}
\usage{
create_summary_table(
  subset_df,
  normalized_counts,
  samp_sheet,
  target_patient,
  target_genes
)
}
\arguments{
\item{subset_df}{A dataframe containing the geoscore of the targeted patient}

\item{normalized_counts}{A matrix of normalized counts. Gene names and sample names must be included.}

\item{samp_sheet}{A sample sheet which associates samples names to groups}

\item{target_patient}{Character indicating which patient the table should be made for}

\item{target_genes}{The genes which will be shown in the table along with the geoscore}
}
\value{
A dataframe with the geoscore and normalized gene counts for the defined patient and genes.
}
\description{
Function which creates a table which summarises the results for a specific patient.
The table is intended to be plotted alongside a plot for that same patient.
}
\author{
Yohan Lefol
}
