% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{expQC}
\alias{expQC}
\title{Expression QC data}
\format{
A data frame with 561 rows and 23 columns.
}
\description{
Quality control metrics for the five cohorts analyzed in NanoString
experiments.
}
\details{
The total number of samples from the five cohorts is 561.
}
\seealso{
\link{cohort}
}
