% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanoString-package.r
\docType{package}
\name{NanoString}
\alias{NanoString}
\alias{NanoString-package}
\title{NanoString: Quality Control for NanoString platform}
\description{
The NanoString package provides a novel set of tools used for quality control purposes
in the NanoString platform. It includes batch adjustment using the ratio method thus far.
}

