% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanostringr-package.r
\docType{package}
\name{nanostringr}
\alias{nanostringr}
\alias{nanostringr-package}
\title{nanostringr: Quality Control and normalization for nanostringr platform}
\description{
The nanostringr package provides a novel set of tools used for quality control purposes
in the NanoString platform. It includes batch adjustment using the ratio method thus far.
}

