% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanostring_YL_functions.R
\name{pagebreak}
\alias{pagebreak}
\title{pagebreak}
\usage{
pagebreak()
}
\value{
The correct command for a pagebreak (latex/PDF or html)
}
\description{
A simple function to create a pagebreak in a rmarkdown report.
}
\author{
Yohan Lefol
}
