% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc_to_codeset.R
\name{rcc_to_codeset}
\alias{rcc_to_codeset}
\title{Convert directory of RCC files to a codeset}
\usage{
rcc_to_codeset(dir)
}
\arguments{
\item{dir}{directory path where RCC files exist}
}
\value{
A tibble with columns "Code.Class", "Name", "Accession", and counts
  from each sample.
}
\description{
Reads in RCC files from a directory and reformats to a codeset that can be
used for analysis
}
