% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rcc.R
\name{read_rcc}
\alias{read_rcc}
\title{Read NanoString RCC files}
\usage{
read_rcc(path = ".", pattern = "\\\\.RCC$")
}
\arguments{
\item{pattern}{file pattern for matching file extensions ".RCC" (or ".rcc")}

\item{dir}{directory path where RCC files exist}
}
\value{
A tibble with columns "File.Name", "fov.count", "fov.counted",
  "binding.density".
}
\description{
Read a directory of RCC files and merge into count data and sample
attributes.
}
\details{
RCC files for a sample are direct outputs from NanoString runs. We can
extract counts for each gene in a sample. Sample attributes include
information about Fov count, Fov counted, binding density, date, GeneRLF,
SampleID, etc. Both count and attribute data are then merged across samples.
}
