% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanostringr-package.r
\docType{package}
\name{nanostringr-package}
\alias{nanostringr}
\alias{nanostringr-package}
\title{nanostringr: Performs Quality Control, Data Normalization, and Batch
    Effect Correction for NanoString nCounter Data}
\description{
QC, normalization, and batch adjustment operations for NanoString
    data. Raw NanoString data can be imported in the form of RCC files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OVCARE/nanostringr}
  \item Report bugs at \url{https://github.com/OVCARE/nanostringr/issues}
}

}
\author{
\strong{Maintainer}: Derek Chiu \email{dchiu@bccrc.ca}

Authors:
\itemize{
  \item Aline Talhouk \email{atalhouk@bccrc.ca}
  \item Samuel Leung \email{Samuel.Leung@vch.ca}
}

}
\keyword{internal}
