% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanostringr-package.r
\docType{package}
\name{nanostringr-package}
\alias{nanostringr}
\alias{nanostringr-package}
\title{nanostringr: Performs Quality Control and Data Normalization and Batch Effect Correction for NanoString nCounter Data}
\description{
QC and batch adjustment operations for NanoString data.
}
\author{
\strong{Maintainer}: Derek Chiu \email{dchiu@bccrc.ca}

Authors:
\itemize{
  \item Aline Talhouk \email{atalhouk@bccrc.ca}
  \item Samuel Leung \email{Samuel.Leung@vch.ca}
}

}
\keyword{internal}
