% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanostringr-package.R
\docType{package}
\name{nanostringr-package}
\alias{nanostringr}
\alias{nanostringr-package}
\title{nanostringr: Performs Quality Control, Data Normalization, and Batch Effect Correction for 'NanoString nCounter' Data}
\description{
Provides quality control (QC), normalization, and batch effect correction operations for 'NanoString nCounter' data, Talhouk et al. (2016) \doi{10.1371/journal.pone.0153844}. Various metrics are used to determine which samples passed or failed QC. Gene expression should first be normalized to housekeeping genes, before a reference-based approach is used to adjust for batch effects. Raw NanoString data can be imported in the form of Reporter Code Count (RCC) files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TalhoukLab/nanostringr/}
  \item \url{https://talhouklab.github.io/nanostringr/}
  \item Report bugs at \url{https://github.com/TalhoukLab/nanostringr/issues}
}

}
\author{
\strong{Maintainer}: Derek Chiu \email{dchiu@bccrc.ca} (\href{https://orcid.org/0000-0002-7591-4881}{ORCID})

Authors:
\itemize{
  \item Aline Talhouk \email{atalhouk@bccrc.ca} (\href{https://orcid.org/0000-0001-7760-410X}{ORCID})
  \item Samuel Leung \email{Samuel.Leung@vch.ca} (\href{https://orcid.org/0000-0003-0138-7254}{ORCID})
}

}
\keyword{internal}
