% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NanostringR-package.R
\docType{package}
\name{NanostringR-package}
\alias{NanostringR}
\alias{NanostringR-package}
\title{NanostringR: Quality Control and Normalization for NanoString Gene Expression Data}
\description{
This package consolidates two unmaintained packages --- NanoStringNorm
(Waggott et al., 2012) and NanoStringQCPro (Bioconductor) --- with custom
functions for clinical NanoString analysis workflows. The source code of those
packages is included verbatim; all credit for those contributions belongs to
their original authors.
}
\author{
\strong{Maintainer}: Yohan Lefol \email{yohan.lefol@gmail.com}

}
\keyword{internal}
