% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rcc.R
\name{rcc}
\alias{rcc}
\alias{read_rcc}
\alias{parse_counts}
\alias{parse_attributes}
\title{Read NanoString RCC files}
\usage{
read_rcc(path = ".", pattern = "\\\\.RCC$")

parse_counts(file)

parse_attributes(file)
}
\arguments{
\item{path}{directory path for multiple RCC files}

\item{pattern}{pattern for matching file extensions ".RCC" or ".rcc"}

\item{file}{RCC file name}
}
\value{
A tibble with columns "File.Name", "fov.count", "fov.counted",
  "binding.density".
}
\description{
Read directory of RCC files and merge samples into count and attribute
data.
}
\details{
RCC files for a sample are direct outputs from NanoString runs. We can
extract counts for each gene in a sample. Sample attributes include sample
ID, GeneRLF, date, cartridge ID, lane number, Fov count, Fov counted, and
binding density. Both count and attribute data are then merged across
samples.

If `path` points to a zipped RCC file with multiple samples, the zip file is
unzipped and a directory of RCC sample files are created with the same name.
}
\author{
Derek Chiu
}
