% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcc_to_exp.R
\name{rcc_to_exp}
\alias{rcc_to_exp}
\title{Convert directory of RCC files to an expression object}
\usage{
rcc_to_exp(dir)
}
\arguments{
\item{dir}{directory path where RCC files exist}
}
\value{
A tibble with columns "File.Name", "fov.count", "fov.counted",
  "binding.density".
}
\description{
Reads in RCC files from a directory and reformats to an expression object
used for analysis
}
